Added initial implementation of Residual Att Net + notebook to test it
Implemented SoftMask branch and AttentionModule classes.
Not sure if attention module is correct, as paper is quite cryptic.
Notebook contains an initial implementation of Attention56, which is
supposed to be used in 224x224 images.
Current version seems not to be "learning" anything, but may be effect
of small dataset size